requirements.txt
streamlit
groq
googlesearch-python
requests
beautifulsoup4